Replace WeakSet with set for event subscribers
The use of WeakSet for subscriber storage has been replaced with a regular set to simplify implementation. This change eliminates the need for weak references in this context, as no specific weak reference behavior is required. Weak set was not allowing acces to the set after the method has been created